Move tunnel types file to types folder
