+++ webapp/app/[locale]/tunnel/_components/prove.tsx
@@ -1,10 +1,10 @@
 import { MessageStatus } from '@eth-optimism/sdk'
-import { EvmWithdrawOperation } from 'app/context/tunnelHistoryContext/types'
 import { evmRemoteNetworks, hemi } from 'app/networks'
 import { useChain } from 'hooks/useChain'
 import { useTunnelHistory } from 'hooks/useTunnelHistory'
 import { useTranslations } from 'next-intl'
 import { useEffect, useState } from 'react'
+import { EvmWithdrawOperation } from 'types/tunnel'
 import { Button } from 'ui-common/components/button'
 import { formatNumber } from 'utils/format'
 import { getL2TokenByBridgedAddress, getTokenByAddress } from 'utils/token'
