+++ webapp/app/[locale]/tunnel/_components/reviewOperation/reviewBtcDeposit.tsx
@@ -1,10 +1,10 @@
 import { RemoteChain } from 'app/networks'
 import { TransactionStatus } from 'components/transactionStatus'
-import { BtcDepositStatus } from 'context/tunnelHistoryContext/types'
 import { useBtcDeposits } from 'hooks/useBtcDeposits'
 import { useTranslations } from 'next-intl'
 import { type ComponentProps, type FormEvent, type ReactNode } from 'react'
 import { Token } from 'types/token'
+import { BtcDepositStatus } from 'types/tunnel'
 import { Card } from 'ui-common/components/card'
 import { CloseIcon } from 'ui-common/components/closeIcon'
 import { Modal } from 'ui-common/components/modal'
