+++ webapp/app/[locale]/tunnel/_components/reviewOperation/reviewEvmWithdrawal.tsx
@@ -1,11 +1,11 @@
 import { MessageStatus } from '@eth-optimism/sdk'
 import { evmRemoteNetworks, hemi } from 'app/networks'
 import { TransactionStatus } from 'components/transactionStatus'
-import { EvmWithdrawOperation } from 'context/tunnelHistoryContext/types'
 import { useTunnelHistory } from 'hooks/useTunnelHistory'
 import { useTranslations } from 'next-intl'
 import { FormEvent, ReactNode } from 'react'
 import Skeleton from 'react-loading-skeleton'
+import { EvmWithdrawOperation } from 'types/tunnel'
 import { Card } from 'ui-common/components/card'
 import { Modal } from 'ui-common/components/modal'
 import { getL2TokenByBridgedAddress, getTokenByAddress } from 'utils/token'
